requirements.txt
fastapi
uvicorn
yt-dlp
demucs
openai-whisper
python-docx